+++ src/CiDetector.php
@@ -16,8 +16,8 @@ class CiDetector
     const CI_CODESHIP = 'Codeship';
     const CI_CONTINUOUSPHP = 'continuousphp';
     const CI_DRONE = 'drone';
-    const CI_GITLAB = 'GitLab';
     const CI_GITHUB_ACTIONS = 'GitHub Actions';
+    const CI_GITLAB = 'GitLab';
     const CI_JENKINS = 'Jenkins';
     const CI_TEAMCITY = 'TeamCity';
     const CI_TRAVIS = 'Travis CI';
@@ -60,8 +60,8 @@ class CiDetector
             Ci\Codeship::class,
             Ci\Continuousphp::class,
             Ci\Drone::class,
-            Ci\GitLab::class,
             Ci\GitHubActions::class,
+            Ci\GitLab::class,
             Ci\Jenkins::class,
             Ci\TeamCity::class,
             Ci\Travis::class,